Update of all the work done
Includes:
P,I,D,PT1,PT2,PID blocks
contains a .xls with graphs showcasing a test run of all blocks avaliable
_!WARNING!!_
Contains a PTt block which is, currrently, useless. It should work, but doesnt. It is yet to be determined if the fault lies in the user OR in the mathematics behind it.